getting all ready
