+++ crates/cheatcodes/Cargo.toml
@@ -41,9 +41,8 @@ alloy-signer-local = { workspace = true, features = [
 ] }
 parking_lot.workspace = true
 alloy-network.workspace = true
-# NOTE(zk): dependency moved to strategy crate would
-# alloy-consensus = { workspace = true, features = ["k256"] }
-
+alloy-consensus = { workspace = true, features = ["k256"] }
+alloy-rlp.workspace = true
 
 base64.workspace = true
 dialoguer = "0.11"
