+++ crates/cheatcodes/src/evm.rs
@@ -5,8 +5,10 @@ use crate::{
     BroadcastableTransaction, Cheatcode, Cheatcodes, CheatcodesExecutor, CheatsCtxt, Error, Result,
     Vm::*,
 };
+use alloy_consensus::TxEnvelope;
 use alloy_genesis::{Genesis, GenesisAccount};
 use alloy_primitives::{Address, Bytes, B256, U256};
+use alloy_rlp::Decodable;
 use alloy_sol_types::SolValue;
 use foundry_common::fs::{read_json_file, write_json_file};
 use foundry_evm_core::{
@@ -767,23 +769,29 @@ impl Cheatcode for getStateDiffJsonCall {
 
 impl Cheatcode for broadcastRawTransactionCall {
     fn apply_full(&self, ccx: &mut CheatsCtxt, executor: &mut dyn CheatcodesExecutor) -> Result {
-        // Note(zk): The function transact_from_tx has a different signature compered to upstream.
-        // It was changed to allow passing EIP-712 data for zkSync. This is handled inside the
-        // strategy. We are now delegating the tx decoding step to the runner (which know
-        // which strategy to use)
-        let tx = ccx.ecx.db.transact_from_tx(
+        // let tx = TxEnvelope::decode(&mut self.data.as_ref())
+        //     .map_err(|err| fmt_err!("failed to decode RLP-encoded transaction: {err}"))?;
+
+        // ccx.ecx.db.transact_from_tx(
+        //     &tx.clone().into(),
+        //     (*ccx.ecx.env).clone(),
+        //     &mut ccx.ecx.journaled_state,
+        //     &mut *executor.get_inspector(ccx.state),
+        // )?;
+
+        // if ccx.state.broadcast.is_some() {
+        //     ccx.state.broadcastable_transactions.push_back(BroadcastableTransaction {
+        //         rpc: ccx.db.active_fork_url(),
+        //         transaction: tx.try_into()?,
+        //     });
+        // }
+
+        ccx.ecx.db.transact_from_tx_zk(
             &self.data,
             (*ccx.ecx.env).clone(),
             &mut ccx.ecx.journaled_state,
             &mut *executor.get_inspector(ccx.state),
         )?;
-
-        if ccx.state.broadcast.is_some() {
-            ccx.state.broadcastable_transactions.push_back(BroadcastableTransaction {
-                rpc: ccx.db.active_fork_url(),
-                transaction: tx,
-            });
-        }
 
         Ok(Default::default())
     }
